feat: 멤버 엔티티 필드 수정
