--- src/main/java/org/sopt/domain/Member.java
@@ -1,20 +1,0 @@
-package org.sopt.domain;
-
-public class Member {
-
-    private Long id;
-    private String name;
-
-    public Member(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
